feat: added SideNavigation.vue
