+++ src/main.js
@@ -28,7 +28,6 @@ const vuetify = createVuetify({
       dark: {
         primary: '#BB86FC',
         secondary: '#03DAC6',
-        error: '#D32F2F',
         // Define other dark theme colors as needed
       },
       light: {
